requirements.txt
flask==2.3.2
python-telegram-bot==20.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.96.0
requests==2.31.0
python-dotenv==1.0.0
